Add metrics.collectors()
metrics.collect() exposes observations as single
array. So there's no way to distinguish histograms
from a set of metrics.

E.g. in prometheus we need to add metainfo
on the histogram type when exporting metrics.

metrics.collect() now returns an array of
collectors. Which in turn have :collect() method
to return current observations.