bugfix: psutils cpu count after clear
Before this patch, tnt_cpu_count metric collector was built on psutils
script require. Calling metrics.clear() removes tnt_cpu_count metric
from the registry, so the only way to restore it after is to manually
reload psutils script. This patch moves collector setup to update
handle, so it would be rebuilt if needed.

After this patch it is also possible to load metrics.psutils.cpu package
without creating a collector in the global registry. It is
a prerequirement for embedding the module to the core tarantool.

Part of tarantool/tarantool#7725