api: allow to invoke callbacks before collect
Before this patch, any user that wanted to export metrics to their Lua
code needed to perform a consecutive call of `metrics.invoke_callbacks`
and `metrics.collect`. Now it could be done with a single `collect` call
with `invoke_callbacks` option set to `true`.

Part of tarantool/tarantool#8192, part of tarantool/tarantool#7725