api: allow to collect only default metrics
After this patch, it is possible to collect only default metrics in Lua.
Collector metainfo is used to decide which metric is default and which
is not. Beware that collect does not control the list of enabled metrics
-- if some (or all) default metrics were disabled or weren't enabled,
they won't be appear in observation.

The motivation is a requirement from Tarantool Feedback Daemon [1] to
collect default metrics, but not the user ones.

1. https://github.com/tarantool/tarantool/issues/8192

Part of tarantool/tarantool#8192, part of tarantool/tarantool#7725