api: set version in code
Before this patch, it was assumed that version would be set
automatically with CI. Yet, current approach don't work. This patch
replaces this approach with simple manual solution.

The rules are as follows:
- bump version to X.Y.Z in release commit,
- bump version to X.Y.Z-scm in the following commit.

luarocks.core.vers comparison treats scm version as we expect them to:
- X.Y.Z-scm > X.Y.Z
- X.Y.Z+1 > X.Y.Z-scm
- X.Y+1.0 > X.Y.Z-scm
- X+1.0.0 > X.Y.Z-scm

The version approach may be reworked later, but since we need a valid
version info for tarantool/tarantool#8192, we implement this one at
least for now.

Closes #330, part of tarantool/tarantool#8192