api: deprecate using {} to include all metrics
Before this series of patches, `{}` (empty table) were equivalent
to new `'all'`. It may be counter-intuitive since `include` contains
a list of metric classes to be enabled. It is also still possible
to pass `nil` to use the default behavior as before, the only thing that
is deprecated is an explicit empty table.

Part of tarantool/tarantool#7725