cartridge: setup hotreload inside the role
Current cartridge hotreload is based on the following assumption:
- `cartridge.hotreload` is loaded before metrics package.

It's not always true, so package.loaded check should be replaced with
require. On the other hand,
- being able to require `cartridge.hotreload` not means that
  the application is a cartridge one.

For example, rock may be installed globally on a server with multiple
applications, but application with metrics may be a non-cartridge one.

It seems that it is better to move hotreload process to cartridge role
initialization, since using a role impossible in a non-cartridge
application. crud module uses the same approach [1].

1. https://github.com/tarantool/crud/commit/6da4f5684a00fe39106ac139538b06821c8c829c

Part of tarantool/tarantool#7725