api: introduce collector name prefixes
Collector name prefixes may be used to build names for metric
derivatives (average, min, max etc.). For now the rule is as follows:
name prefix is the name stripped from the "_total" and "_current"
suffixes. (If they presents together at the same time, the name is
already messed up, so we don't cover this.)

Part of tarantool/tarantool#7725
Part of tarantool/tarantool#7728