rock: build with cmake
The ability to override built-in modules was introduced with [1]. To use
it, you must install the package files using the override.package path
for each file included.

Using soft links inside the rock seems like a bad idea -- luarocks try
to resolve them on files deploy and fails to do so. The only stable
approach is seems to be to duplicate the files to the override
folder.

builtin rockspec doesn't allow to specify a single file twice, so we
need to store the soft link in the repo to workaround it. To build
override package with builtin rockspec, we'll also need to manually
maintain two lists of package files.

cmake approach allows to automatize it. On the other hand, to work with
cmake-built rock, you'll need to `make` the rock while working locally.
Since `make .rocks` is a dependency of `make test`, it shouldn't be
a big issue.

1. https://github.com/tarantool/tarantool/issues/7774

Part of tarantool/tarantool#7727