rock: override built-in metrics
The ability to override built-in modules was introduced with [1].
After [2], metrics will be embedded to core Tarantool. The ability to
override them with installed rock will make it possible to use old
Tarantool with new metrics. cartridge role, which was embedded to
tarantool/cartridge in [3], is also would be override.

1. https://github.com/tarantool/tarantool/issues/7774
2. https://github.com/tarantool/tarantool/issues/7725
3. https://github.com/tarantool/cartridge/pull/2047

Closes tarantool/tarantool#7727