test: workaround LUA_PATH only for repo tests
luatest implicitly requires `test.helper` package in the beginning of a
test run, if there is a package. We use the assumption that
`test.helper` package exist and required to find the module root in
`cfg_test` in tarantool/metrics.

These tests are also included in tarantool/tarantool repo as
a submodule. Paths are workarounded before the run [1] so we can reuse
existing tests and not rewrite them. But `test.helper` path is not
workarounded since we do not need this file. Moreover, we shouldn't use
it in tarantool/tarantool since it asserts that non-built-in package is
used, while we test built-in one there. But tarantool/tarantool tests
actually use `test.helper` to find the root in `cfg_test`. It had worked
for tarantool/tarantool tests before because, in fact, we had used
luatest `test.helper` package to find the root (and, for some reason, it
was fine). After [2], there is no luatest `test.helper` anymore and root
search fails [3].

This patch makes `cfg_test` to set LUA_PATH to `nil` if we're testing
built-in package in tarantool/tarantool since we don't need any rock
files.

1. https://github.com/tarantool/tarantool/blob/13159230cfd9ec46b93ad7e6185a3204e27b5b50/test/metrics-luatest/helper.lua#L111-L112
2. https://github.com/tarantool/luatest/pull/307
3. https://github.com/tarantool/tarantool/pull/8683#issuecomment-1558074337

Closes #456