Bump cmake version. Bump ubuntu version in the CI. (#505)
* ci: update ubuntu version

The `ubuntu-20.04 (focal)` image is no longer supported, so the build
using this image returns the following error:

| This is a scheduled Ubuntu 20.04 brownout. Ubuntu 20.04 LTS runner
| will be removed on 2025-04-01.

See [1] for the details.

This patch bumps the used version to the ubuntu-24.04 (noble).

[1]: https://github.com/actions/runner-images/issues/11101

Closes #504
Needed for tarantool/tarantool#11220

* ci: remove EOL Tarantool versions from testing

This patch drops EOL Tarantool versions that have no package available
for Ubuntu 24.04 noble at download.tarantool.org.

* cmake: update the minimum required version

For CMake 4.0, compatibility with versions of CMake older than 3.5 has
been removed [1]. Thus, on macOS it leads to the error on build, since
Homebrew updates it to version 4.0.

This patch sets the minimum required version to 3.5 for now.

[1]: https://cmake.org/cmake/help/latest/release/4.0.html#deprecated-and-removed-features